Change the begin method to return an uint8_t instead of an int. The code before returned an int on basically a bool. This will save 8 bits of RAM, right?

+++ include/pcf8563.h
@@ -114,7 +114,7 @@ class RTC_Alarm
 class PCF8563_Class
 {
     public:
-        int begin(TwoWire &port = Wire, uint8_t addr = PCF8563_SLAVE_ADDRESS);
+        uint8_t begin(TwoWire &port = Wire, uint8_t addr = PCF8563_SLAVE_ADDRESS);
         void check();
         void setDateTime(
             uint16_t year,
